fix(worker): thin __main__ to stop spawn re-importing heavy worker closure (Fix A)
Moves all worker logic from src/worker/main.py into the new src/worker/app.py.
Rewrites main.py to be stdlib-only at module level, breaking the spawn
re-import path that was inflating the template process from ~97MB to ~210MB.

After this change the template process (multiprocessing spawn_main) drops
from ~206MB → ~111MB RSS on the debug stack, within the plan's ~95-110MB
target. Worker container total drops from ~390MiB → ~302MiB.

Also deletes the dead `preload_modules` parameter from _template_main in
template_process.py — it was never passed by any caller (no-dead-code rule).

Test results:
- test_worker_spawn_entry_is_stdlib_thin: GREEN (was red)
- test_template_process_module_is_stdlib_thin: GREEN (stays green)
- test_worker_app_closure_has_no_heavyweights: still red (expected — Tasks 3-6)
- test_scheduler_closure_has_no_heavyweights: still red (expected — Task 3)
- ./test.sh unit: 4057 passed, 2 expected-red failures
- ./test.sh e2e tests/e2e/engine/: 9/9 passed
- pyright src/worker/ + template_process.py: 0 errors
- ruff: clean
